Поправила структуру
